# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Job matching API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs/{id}/run-matching:
    post:
      operationId: runJobMatching
      tags: [matches]
      summary: Run AI matching for a job against all active candidates
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Matching results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"
  /jobs/{id}/matches:
    get:
      operationId: getJobMatches
      tags: [matches]
      summary: Get all matches for a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of matches for the job
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchWithCandidate"
  /candidates/{id}/matches:
    get:
      operationId: getCandidateMatches
      tags: [matches]
      summary: Get all matches for a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of matches for the candidate
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchWithJob"
  /matches/{id}:
    patch:
      operationId: updateMatchStatus
      tags: [matches]
      summary: Update match status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMatchStatusBody"
      responses:
        "200":
          description: Match updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
...
  /dashboard/recent-matches:
    get:
      operationId: getRecentMatches
      tags: [dashboard]
      summary: Get recent match results
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecentMatch"
  /dashboard/top-candidates:
    get:
      operationId: getTopCandidates
      tags: [dashboard]
      summary: Get top matched candidates across all jobs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Top candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopCandidate"
...
components:
  schemas:
...
    Match:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        candidateId:
          type: integer
        overallScore:
          type: number
        skillScore:
          type: number
        experienceScore:
          type: number
        educationScore:
          type: number
        locationScore:
          type: number
        assessment:
          type: string
        matchedSkills:
          type: array
          items:
            type: string
        missingSkills:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, shortlisted, rejected, hired]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - jobId
        - candidateId
        - overallScore
        - skillScore
        - experienceScore
        - educationScore
        - locationScore
        - assessment
        - matchedSkills
        - missingSkills
        - status
        - createdAt
    MatchWithCandidate:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        candidateId:
          type: integer
        overallScore:
          type: number
        skillScore:
          type: number
        experienceScore:
          type: number
        educationScore:
          type: number
        locationScore:
          type: number
        assessment:
          type: string
        matchedSkills:
          type: array
          items:
            type: string
        missingSkills:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, shortlisted, rejected, hired]
        createdAt:
          type: string
          format: date-time
        candidateName:
          type: string
        candidateTitle:
          type: string
        candidateEmail:
          type: string
      required:
        - id
        - jobId
        - candidateId
        - overallScore
        - skillScore
        - experienceScore
        - educationScore
        - locationScore
        - assessment
        - matchedSkills
        - missingSkills
        - status
        - createdAt
        - candidateName
        - candidateTitle
        - candidateEmail
    MatchWithJob:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        candidateId:
          type: integer
        overallScore:
          type: number
        skillScore:
          type: number
        experienceScore:
          type: number
        educationScore:
          type: number
        locationScore:
          type: number
        assessment:
          type: string
        matchedSkills:
          type: array
          items:
            type: string
        missingSkills:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, shortlisted, rejected, hired]
        createdAt:
          type: string
          format: date-time
        jobTitle:
          type: string
        jobCompany:
          type: string
      required:
        - id
        - jobId
        - candidateId
        - overallScore
        - skillScore
        - experienceScore
        - educationScore
        - locationScore
        - assessment
        - matchedSkills
        - missingSkills
        - status
        - createdAt
        - jobTitle
        - jobCompany
    UpdateMatchStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, shortlisted, rejected, hired]
      required:
        - status
...
    RecentMatch:
      type: object
      properties:
        id:
          type: integer
        jobTitle:
          type: string
        candidateName:
          type: string
        overallScore:
          type: number
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - jobTitle
        - candidateName
        - overallScore
        - status
        - createdAt
    TopCandidate:
      type: object
      properties:
        candidateId:
          type: integer
        candidateName:
          type: string
        candidateTitle:
          type: string
        avgScore:
          type: number
        matchCount:
          type: integer
        topSkills:
          type: array
          items:
            type: string
      required:
        - candidateId
        - candidateName
        - candidateTitle
        - avgScore
        - matchCount
        - topSkills
...