# lib/api-zod/src/generated/types/matchWithJob.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job matching API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MatchWithJobStatus } from "./matchWithJobStatus";

export interface MatchWithJob {
  id: number;
  jobId: number;
  candidateId: number;
  overallScore: number;
  skillScore: number;
  experienceScore: number;
  educationScore: number;
  locationScore: number;
  assessment: string;
  matchedSkills: string[];
  missingSkills: string[];
  status: MatchWithJobStatus;
  createdAt: Date;
  jobTitle: string;
  jobCompany: string;
}