# lib/api-zod/src/generated/types/recentMatch.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job matching API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RecentMatch {
  id: number;
  jobTitle: string;
  candidateName: string;
  overallScore: number;
  status: string;
  createdAt: Date;
}