# lib/api-zod/src/generated/types/updateMatchStatusBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job matching API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateMatchStatusBodyStatus } from "./updateMatchStatusBodyStatus";

export interface UpdateMatchStatusBody {
  status: UpdateMatchStatusBodyStatus;
}